- Added script to perform Sobel Edge detection without the use of OpenCV - Updated README with image after application of Canny filter
